Factor out fixtures common to app/generate tests
