+++ tests/test_app/test_generate_vegattributes_json.py
@@ -1,28 +1,11 @@
-from dataclasses import dataclass
 import json
 from pathlib import Path
-import pytest
 from unittest.mock import patch
-from typing import List
 
-from fooddata_vegattributes.fooddata import FoodDataDict
 from fooddata_vegattributes.app.generate import main
 
+from .conftest import FakeFoodDataJson
 
-@dataclass
-class FakeFoodDataJson:
-  food_data_dicts: List[FoodDataDict]
-  path: Path
-
-@pytest.fixture
-def fake_fooddata_json(tmp_path: Path):
-  json_path = tmp_path/"fake_fooddata.json"
-  food_data = [{ "fdcId": 123456, "description": "A fake food, salted" }]
-
-  with json_path.open("w") as f:
-    json.dump({ "SurveyFoods": food_data }, f)
-
-  yield FakeFoodDataJson(food_data_dicts=food_data, path=json_path)
 
 def test_generate_vegattributes_json(
   fake_fooddata_json: FakeFoodDataJson,
